clean pages.
