+++ src/App.tsx
@@ -3,11 +3,9 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
 import loadGoogleMapsAPI from "./utils/gMap";
 
-// Component
 import Header from "./components/global/Header/Header";
 import Footer from "./components/global/Footer/Footer";
 
-// Pages
 import HomePage from "./pages/HomePage/HomePage";
 import Register from "./pages/Register/Register";
 import RegisterFormPage from "./pages/RegisterFormPage/RegisterFormPage";
@@ -15,11 +13,9 @@ import DetailPage from "./pages/DetailPage/DetailPage";
 import DevPage from "./pages/DevPage/DevPage";
 import ExplorePage from "./pages/ExplorePage/ExplorePage";
 
-// Apollo Client
 import client from "./graphql/apolloClient";
 import { ApolloProvider } from "@apollo/client";
 
-// Google Maps Context
 const GoogleMapsContext = React.createContext<any>(null);
 
 function GoogleMapsProvider({ children }: { children: React.ReactNode }) {
@@ -56,10 +52,8 @@ function App() {
     <ApolloProvider client={client}>
       <GoogleMapsProvider>
         <Router>
-          {/* Header component */}
           <Header />
 
-          {/* Define routes */}
           <Routes>
             <Route path="/" element={<HomePage />} />
             <Route path="/register" element={<Register />} />
@@ -69,7 +63,6 @@ function App() {
             <Route path="/detail" element={<DetailPage />} />
           </Routes>
 
-          {/* Footer component */}
           <Footer />
         </Router>
       </GoogleMapsProvider>
